# swagger/api.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API"
  },
  "host": "momo-wars-dev.herokuapp.com",
  "basePath": "/api",
  "tags": [
    {
      "name": "Accounts"
    },
    {
      "name": "Maps"
    },
    {
      "name": "Matches"
    },
    {
      "name": "Players"
    },
    {
      "name": "Sessions"
    },
    {
      "name": "Units"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/v1/maps": {
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "GetMaps",
        "description": "Returns the list of available maps",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Map"
              }
            },
            "description": "Successful operation"
          },
          "401": {
            "description": "Session not valid"
          }
        }
      }
    },
    "/v1/account": {
      "patch": {
        "tags": [
          "Accounts"
        ],
        "summary": "UpdateAccount",
        "description": "Updates a user's account",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.endpoints.accounts.UpdateAccount.DataParameter"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Account"
            },
            "description": "Successful operation"
          },
          "400": {
            "description": "Parameters missing or invalid"
          },
          "401": {
            "description": "Session not valid"
          }
        }
      },
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "CreateAccount",
        "description": "Creates a user's account",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.endpoints.accounts.CreateAccount.DataParameter"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Account"
            },
            "description": "Successful operation"
          },
          "400": {
            "description": "Parameters missing or invalid"
          },
          "409": {
            "description": "An account with the same email already exists"
          }
        }
      },
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "GetAccount",
        "description": "Returns the user's account",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Account"
            },
            "description": "Successful operation"
          },
          "401": {
            "description": "Session not valid"
          }
        }
      }
    },
    "/v1/players/{playerId}/turn": {
      "delete": {
        "tags": [
          "Players"
        ],
        "summary": "EndTurn",
        "description": "Ends the turn of the player for the given match",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Parameters missing or invalid"
          },
          "422": {
            "description": "Not allowed to end the turn on the given match"
          },
          "401": {
            "description": "Session not valid"
          },
          "204": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Player not found"
          }
        }
      }
    },
    "/v1/players/{playerId}": {
      "delete": {
        "tags": [
          "Players"
        ],
        "summary": "LeaveMatch",
        "description": "Leaves the current player's match",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Parameters missing or invalid"
          },
          "422": {
            "description": "Player already left the match or the match is finished"
          },
          "401": {
            "description": "Session not valid"
          },
          "204": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/v1/maps/{mapId}": {
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "GetMap",
        "description": "Returns a map by its id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "mapId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Map"
            },
            "description": "Successful operation"
          },
          "401": {
            "description": "Session not valid"
          },
          "404": {
            "description": "Invalid map id"
          }
        }
      }
    },
    "/v1/session/email": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "CreateSessionEmail",
        "description": "Creates a new session with email and password",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.endpoints.sessions.CreateSessionEmail.DataParameter"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Parameters missing or invalid"
          },
          "401": {
            "description": "Invalid email or password"
          }
        }
      }
    },
    "/v1/matches/{matchId}/players/{playerId}/units/{unitId}/move": {
      "put": {
        "tags": [
          "Units"
        ],
        "summary": "MoveUnits",
        "description": "Enqueues the command to move the units",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "cellId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "unitId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.endpoints.units.MoveUnits.DataParameter"
            },
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Parameters missing or invalid"
          },
          "422": {
            "description": "Invalid matchId, cellId or unitId"
          },
          "401": {
            "description": "Session not valid"
          },
          "204": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/v1/matches": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "CreateMatch",
        "description": "Creates a new match",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.endpoints.matches.CreateMatch.DataParameter"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Match"
            },
            "description": "Successful operation"
          },
          "400": {
            "description": "Parameters missing or invalid"
          },
          "422": {
            "description": "The map id doesn't correspond to a valid map"
          },
          "401": {
            "description": "Session not valid"
          }
        }
      },
      "get": {
        "tags": [
          "Matches"
        ],
        "summary": "GetOpenMatches",
        "description": "Returns a list of all the open matches",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Match"
              }
            },
            "description": "Successful operation"
          },
          "401": {
            "description": "Session not valid"
          }
        }
      }
    },
    "/v1/matches/{matchId}/players": {
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "JoinMatch",
        "description": "Joins the given match",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Match"
            },
            "description": "Successful operation"
          },
          "422": {
            "description": "Either: the player is already in the match or the match is playing or the match is finished."
          },
          "401": {
            "description": "Session not valid"
          },
          "404": {
            "description": "Invalid match id"
          }
        }
      }
    },
    "/v1/matches/{matchId}/players/{playerId}/cells/{cellId}/queue": {
      "put": {
        "tags": [
          "Units"
        ],
        "summary": "RecruitUnits",
        "description": "Enqueues the command to recruit of the units",
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "cellId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.endpoints.units.RecruitUnits.DataParameter"
            },
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Parameters missing or invalid"
          },
          "422": {
            "description": "Invalid matchId or cellId"
          },
          "401": {
            "description": "Session not valid"
          },
          "204": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/v1/session/google": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "CreateSessionGoogle",
        "description": "Creates a new session with a Google token",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.endpoints.sessions.CreateSessionEmail.DataParameter"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Parameters missing or invalid"
          },
          "401": {
            "description": "Invalid token"
          }
        }
      }
    },
    "/v1/matches/{matchId}": {
      "get": {
        "tags": [
          "Matches"
        ],
        "summary": "GetMatch",
        "description": "Returns a match by its id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Cookie",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Match"
            },
            "description": "Successful operation"
          },
          "401": {
            "description": "Session not valid"
          },
          "403": {
            "description": "Not allowed to retrieve the match"
          }
        }
      }
    }
  },
  "definitions": {
    "com.mauriciotogneri.momowars.api.endpoints.accounts.UpdateAccount.DataParameter": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "nickname": {
          "type": "string",
          "maxLength": 30
        }
      },
      "required": [
        "password",
        "nickname"
      ]
    },
    "com.mauriciotogneri.momowars.api.endpoints.matches.CreateMatch.DataParameter": {
      "type": "object",
      "properties": {
        "mapId": {
          "type": "integer"
        },
        "maxPlayers": {
          "type": "integer",
          "minimum": 2
        }
      },
      "required": [
        "mapId",
        "maxPlayers"
      ]
    },
    "com.mauriciotogneri.momowars.api.endpoints.sessions.CreateSessionEmail.DataParameter": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "maxLength": 50
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "com.mauriciotogneri.momowars.api.endpoints.units.MoveUnits.DataParameter": {
      "type": "object",
      "properties": {
        "movement": {
          "type": "string",
          "enum": [
            "NONE",
            "UP_LEFT",
            "UP",
            "UP_RIGHT",
            "DOWN_RIGHT",
            "DOWN",
            "DOWN_LEFT"
          ]
        }
      },
      "required": [
        "movement"
      ]
    },
    "com.mauriciotogneri.momowars.api.endpoints.accounts.CreateAccount.DataParameter": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "maxLength": 50
        },
        "password": {
          "type": "string"
        },
        "nickname": {
          "type": "string",
          "maxLength": 30
        }
      },
      "required": [
        "email",
        "password",
        "nickname"
      ]
    },
    "com.mauriciotogneri.momowars.api.model.Cell": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "type": {
          "type": "string",
          "enum": [
            "BASE",
            "GRASS",
            "EARTH",
            "MOUNTAIN",
            "FOREST",
            "DESERT",
            "CONCRETE",
            "WATER",
            "ICE"
          ]
        },
        "x": {
          "type": "integer"
        },
        "y": {
          "type": "integer"
        }
      },
      "required": [
        "id",
        "type",
        "x",
        "y"
      ]
    },
    "com.mauriciotogneri.momowars.api.model.Account": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "matches": {
          "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.AccountMatches"
        }
      },
      "required": [
        "email",
        "nickname",
        "matches"
      ]
    },
    "com.mauriciotogneri.momowars.api.model.AccountMatches": {
      "type": "object",
      "properties": {
        "open": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        },
        "playing": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        },
        "finished": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      },
      "required": [
        "open",
        "playing",
        "finished"
      ]
    },
    "com.mauriciotogneri.momowars.api.endpoints.units.RecruitUnits.DataParameter": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "enum": [
            "SQUAD",
            "MORTAR",
            "HUMVEE",
            "LIGHT_TANK",
            "HEAVY_TANK",
            "HELICOPTER"
          ]
        },
        "quantity": {
          "type": "integer",
          "minimum": 1
        }
      },
      "required": [
        "type",
        "quantity"
      ]
    },
    "com.mauriciotogneri.momowars.api.model.Map": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "width": {
          "type": "integer",
          "minimum": 1
        },
        "height": {
          "type": "integer",
          "minimum": 1
        },
        "cells": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Cell"
          },
          "minLength": 1
        }
      },
      "required": [
        "id",
        "name",
        "width",
        "height",
        "cells"
      ]
    },
    "com.mauriciotogneri.momowars.api.model.Match": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "status": {
          "type": "string",
          "enum": [
            "OPEN",
            "PLAYING",
            "CALCULATING",
            "FINISHED"
          ]
        },
        "map": {
          "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Map"
        },
        "maxPlayers": {
          "type": "integer"
        },
        "players": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/com.mauriciotogneri.momowars.api.model.Player"
          }
        }
      },
      "required": [
        "id",
        "status",
        "map",
        "maxPlayers"
      ]
    },
    "com.mauriciotogneri.momowars.api.model.Player": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "nickname": {
          "type": "string"
        },
        "resources": {
          "type": "integer"
        },
        "status": {
          "type": "string",
          "enum": [
            "PLAYING",
            "WAITING",
            "VICTORIOUS",
            "DEFEATED",
            "SURRENDERED"
          ]
        }
      },
      "required": [
        "nickname",
        "status"
      ]
    }
  }
}